C:/Program Files/Git/api/uploads allow file download/removal only to its owners

+++ controllers/upload.controller.js
@@ -5,9 +5,16 @@ const { pathToUploads, pathToThumbnails } = require('../config');
 
 module.exports = {
   getFullSizeImage(req, res, next) {
+    // may contain sensitive data (gps/exif)
     try {
       const { filename } = req.params;
-      res.sendFile(path.join(pathToUploads, filename));
+      const { groups, userId } = res.locals;
+      const image = await imageService.getImage(filename);
+      if (groups.includes(image.groupId)) {
+        res.sendFile(path.join(pathToUploads, filename));
+      } else {
+        throw new CustomError(`User ${userId} is not permitted to download file ${filename}`, 403);
+      } 
     } catch (err) {
       next(err);
     }
@@ -56,11 +63,17 @@ module.exports = {
   async removeImage(req, res, next) {
     try {
       const { filename } = req.params;
-      const removeMainFile = fileService.removeFile(path.join(pathToUploads, filename));
-      const removeThumbnail = fileService.removeFile(path.join(pathToThumbnails, filename));
-      const removeFromDb = imageService.removeImage(filename);
-      await Promise.all([removeMainFile, removeThumbnail, removeFromDb]);
-      res.status(200).json({ message: `${filename} upload removed` });
+      const image = await imageService.getImage(filename);
+      const { userId } = res.locals;
+      if (userId === image.userId) {
+        const removeMainFile = fileService.removeFile(path.join(pathToUploads, filename));
+        const removeThumbnail = fileService.removeFile(path.join(pathToThumbnails, filename));
+        const removeFromDb = imageService.removeImage(filename);
+        await Promise.all([removeMainFile, removeThumbnail, removeFromDb]);
+        res.status(200).json({ message: `${filename} upload removed` });        
+      } else {
+        throw new CustomError(`User ${userId} is not permitted to remove file ${filename}`, 403);
+      }
     } catch (err) {
       next(err);
     }
